+++ app/controllers/redmine_oauth_controller.rb
@@ -27,6 +27,7 @@ class RedmineOauthController < AccountController
 
   def oauth
     session[:back_url] = params[:back_url]
+    session[:autologin] = params[:autologin]
     oauth_csrf_token = generate_csrf_token
     session[:oauth_csrf_token] = oauth_csrf_token
     case Setting.plugin_redmine_oauth[:oauth_name]
@@ -41,6 +42,18 @@ class RedmineOauthController < AccountController
         redirect_uri: oauth_callback_url,
         state: oauth_csrf_token,
         scope: 'read_user'
+      )
+    when 'Google'
+      redirect_to oauth_client.auth_code.authorize_url(
+        redirect_uri: oauth_callback_url,
+        state: oauth_csrf_token,
+        scope: 'profile email'
+      )
+    when 'Keycloak'
+      redirect_to oauth_client.auth_code.authorize_url(
+        redirect_uri: oauth_callback_url,
+        state: oauth_csrf_token,
+        scope: 'openid email'
       )
     when 'Okta'
       redirect_to oauth_client.auth_code.authorize_url(
@@ -72,6 +85,18 @@ class RedmineOauthController < AccountController
       user_info = JSON.parse(userinfo_response.body)
       user_info['login'] = user_info['username']
       email = user_info['email']
+    when 'Google'
+      token = oauth_client.auth_code.get_token(params['code'], redirect_uri: oauth_callback_url)
+      userinfo_response = token.get('https://openidconnect.googleapis.com/v1/userinfo',
+                                    headers: { 'Accept' => 'application/json' })
+      user_info = JSON.parse(userinfo_response.body)
+      user_info['login'] = user_info['email']
+      email = user_info['email']
+    when 'Keycloak'
+      token = oauth_client.auth_code.get_token(params['code'], redirect_uri: oauth_callback_url)
+      user_info = JWT.decode(token.token, nil, false).first
+      user_info['login'] = user_info['preferred_username']
+      email = user_info['email']
     when 'Okta'
       token = oauth_client.auth_code.get_token(params['code'], redirect_uri: oauth_callback_url)
       userinfo_response = token.get(
@@ -98,6 +123,8 @@ class RedmineOauthController < AccountController
   def try_to_login(email, info)
     params['back_url'] = session[:back_url]
     session.delete :back_url
+    params['autologin'] = session[:autologin]
+    session.delete :autologin
     user = User.joins(:email_addresses).where(email_addresses: { address: email }).first
     if user # Existing user
       if user.registered? # Registered
@@ -166,6 +193,22 @@ class RedmineOauthController < AccountController
           site: site,
           authorize_url: '/oauth/authorize',
           token_url: '/oauth/token'
+        )
+      when 'Google'
+        OAuth2::Client.new(
+          Setting.plugin_redmine_oauth[:client_id],
+          Setting.plugin_redmine_oauth[:client_secret],
+          site: site,
+          authorize_url: '/o/oauth2/v2/auth',
+          token_url: 'https://oauth2.googleapis.com/token'
+        )
+      when 'Keycloak'
+        OAuth2::Client.new(
+          Setting.plugin_redmine_oauth[:client_id],
+          Setting.plugin_redmine_oauth[:client_secret],
+          site: site,
+          authorize_url: "/realms/#{Setting.plugin_redmine_oauth[:tenant_id]}/protocol/openid-connect/auth",
+          token_url: "/realms/#{Setting.plugin_redmine_oauth[:tenant_id]}/protocol/openid-connect/token"
         )
       when 'Okta'
         OAuth2::Client.new(