+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog for Redmine OAuth
 ==========================
 
+2.1.4 *2024-01-12*
+------------------
+    
+    Autologin
+    Google OAuth provider
+    Keycloak OAuth provider
+
+
+* New: #22 - About feature requests and providers enhancement
+* New: #21 - Autologin enhancement
+
 2.1.3 *2023-11-20*
 ------------------